refactor

+++ src/main/java/com/example/springapp1/models/entity/Lens.java
@@ -17,11 +17,8 @@ public class Lens {
     String color;
     String cover;
     String design;
-    float index;
+    Double index;
     String filter;
-    int diameter;
+    Integer diameter;
     String diopter;
-
-
-
 }
